+++ Kernel/Containers/array.hpp
@@ -310,7 +310,7 @@ TMPL int hash (array<T> a);
 // cast implicitely.
 inline array<string>&
 operator<< (array<string>& a, char* x) {
-  return a << string (x);
+  return a << as_string (x);
 }
 #endif
