+++ lolly/data/json.hpp
@@ -53,7 +53,7 @@ public:
 
   // primitives constructors
   json (string value) { rep= tm_new<json_rep> (json_tree (value)); }
-  json (const char* value) {
+  template <size_t N> json (const char (&value)[N]) {
     rep= tm_new<json_rep> (json_tree (string (value)));
   }
   json (bool value) {
